+++ src/collective/volto/formsupport/tests/test_send_action_form.py
@@ -171,6 +171,38 @@ class TestMailSend(unittest.TestCase):
             "form-id": {
                 "@type": "form",
                 "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "xxx",
+                        "field_type": "text",
+                    },
+                ],
+            },
+        }
+        transaction.commit()
+
+        response = self.submit_form(
+            data={
+                "from": "john@doe.com",
+                "block_id": "form-id",
+                "data": [{"field_id": "xxx", "label": "foo", "value": "bar"}],
+            },
+        )
+        transaction.commit()
+        res = response.json()
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            res["message"],
+            "Missing required field: subject or from.",
+        )
+
+    def test_email_not_send_if_all_fields_are_not_in_form_schema(
+        self,
+    ):
+        self.document.blocks = {
+            "form-id": {
+                "@type": "form",
+                "send": ["recipient"],
             },
         }
         transaction.commit()
@@ -185,10 +217,53 @@ class TestMailSend(unittest.TestCase):
         transaction.commit()
         res = response.json()
         self.assertEqual(response.status_code, 400)
-        self.assertEqual(
-            res["message"],
-            "Missing required field: subject or from.",
+        self.assertEqual(res["message"], "Empty form data.")
+
+    def test_email_sent_with_only_fields_from_schema(
+        self,
+    ):
+        self.document.blocks = {
+            "form-id": {
+                "@type": "form",
+                "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "xxx",
+                        "field_type": "text",
+                    },
+                ],
+            },
+        }
+        transaction.commit()
+
+        response = self.submit_form(
+            data={
+                "from": "john@doe.com",
+                "block_id": "form-id",
+                "subject": "test subject",
+                "data": [
+                    {"label": "foo", "value": "foo", "field_id": "xxx"},
+                    {"label": "bar", "value": "bar", "field_id": "yyy"},
+                ],
+            },
         )
+        transaction.commit()
+        res = response.json()
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(
+            res, {"data": [{"field_id": "xxx", "label": "foo", "value": "foo"}]}
+        )
+
+        msg = self.mailhost.messages[0]
+        if isinstance(msg, bytes) and bytes is not str:
+            # Python 3 with Products.MailHost 4.10+
+            msg = msg.decode("utf-8")
+        self.assertIn("Subject: test subject", msg)
+        self.assertIn("From: john@doe.com", msg)
+        self.assertIn("To: site_addr@plone.com", msg)
+        self.assertIn("Reply-To: john@doe.com", msg)
+        self.assertIn("<strong>foo:</strong> foo", msg)
+        self.assertNotIn("<strong>bar:</strong> bar", msg)
 
     def test_email_sent_with_site_recipient(
         self,
@@ -197,6 +272,16 @@ class TestMailSend(unittest.TestCase):
             "form-id": {
                 "@type": "form",
                 "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -205,15 +290,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -233,6 +322,16 @@ class TestMailSend(unittest.TestCase):
                 "@type": "form",
                 "send": True,
                 "httpHeaders": [],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -241,15 +340,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -270,6 +373,16 @@ class TestMailSend(unittest.TestCase):
                     "REMOTE_ADDR",
                     "PATH_INFO",
                 ],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -278,15 +391,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
 
         msg = self.mailhost.messages[1]
         if isinstance(msg, bytes) and bytes is not str:
@@ -308,6 +425,16 @@ class TestMailSend(unittest.TestCase):
             "form-id": {
                 "@type": "form",
                 "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -317,15 +444,19 @@ class TestMailSend(unittest.TestCase):
                 "from": "john@doe.com",
                 "to": "to@spam.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -346,6 +477,16 @@ class TestMailSend(unittest.TestCase):
                 "@type": "form",
                 "default_to": "to@block.com",
                 "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -354,15 +495,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -383,6 +528,16 @@ class TestMailSend(unittest.TestCase):
                 "@type": "form",
                 "default_subject": "block subject",
                 "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
             },
         }
         transaction.commit()
@@ -391,15 +546,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "block_id": "form-id",
             },
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -427,6 +586,14 @@ class TestMailSend(unittest.TestCase):
                         "field_type": "from",
                         "use_as_reply_to": True,
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -435,8 +602,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "smith@doe.com"},
                 ],
                 "block_id": "form-id",
@@ -444,7 +615,7 @@ class TestMailSend(unittest.TestCase):
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -472,6 +643,14 @@ class TestMailSend(unittest.TestCase):
                         "field_type": "from",
                         "use_as_bcc": True,
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -480,8 +659,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "smith@doe.com"},
                 ],
                 "block_id": "form-id",
@@ -489,7 +672,7 @@ class TestMailSend(unittest.TestCase):
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(len(self.mailhost.messages), 2)
         msg = self.mailhost.messages[0]
         bcc_msg = self.mailhost.messages[1]
@@ -517,6 +700,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "test",
                         "field_type": "text",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -529,8 +720,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "test", "label": "Test", "value": "test text"},
                 ],
                 "block_id": "form-id",
@@ -539,7 +734,7 @@ class TestMailSend(unittest.TestCase):
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(len(self.mailhost.messages), 1)
 
     def test_send_attachment_validate_size(
@@ -558,6 +753,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "test",
                         "field_type": "text",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -571,8 +774,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "test", "label": "Test", "value": "test text"},
                 ],
                 "block_id": "form-id",
@@ -604,6 +811,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "contact",
                         "field_type": "from",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -612,8 +827,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "smith@doe.com"},
                 ],
                 "block_id": "form-id",
@@ -621,7 +840,7 @@ class TestMailSend(unittest.TestCase):
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -655,6 +874,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "contact",
                         "field_type": "from",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -663,8 +890,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "smith@doe.com"},
                 ],
                 "block_id": "form-id",
@@ -672,7 +903,7 @@ class TestMailSend(unittest.TestCase):
         )
         transaction.commit()
 
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
 
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
@@ -710,7 +941,21 @@ class TestMailSend(unittest.TestCase):
         self,
     ):
         self.document.blocks = {
-            "form-id": {"@type": "form", "send": True, "email_format": "table"},
+            "form-id": {
+                "@type": "form",
+                "send": True,
+                "email_format": "table",
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
+            },
         }
         transaction.commit()
 
@@ -722,15 +967,15 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": message},
-                    {"label": "Name", "value": name},
+                    {"field_id": "message", "label": "Message", "value": message},
+                    {"field_id": "name", "label": "Name", "value": name},
                 ],
                 "subject": subject,
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -772,7 +1017,21 @@ class TestMailSend(unittest.TestCase):
         self,
     ):
         self.document.blocks = {
-            "form-id": {"@type": "form", "send": True, "email_format": "list"},
+            "form-id": {
+                "@type": "form",
+                "send": True,
+                "email_format": "list",
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
+            },
         }
         transaction.commit()
 
@@ -780,15 +1039,19 @@ class TestMailSend(unittest.TestCase):
             data={
                 "from": "john@doe.com",
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "John"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
                 ],
                 "subject": "test subject",
                 "block_id": "form-id",
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -802,13 +1065,27 @@ class TestMailSend(unittest.TestCase):
 
     def test_send_xml(self):
         self.document.blocks = {
-            "form-id": {"@type": "form", "send": True, "attachXml": True},
+            "form-id": {
+                "@type": "form",
+                "send": True,
+                "attachXml": True,
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
+            },
         }
         transaction.commit()
 
         form_data = [
-            {"label": "Message", "value": "just want to say hi"},
-            {"label": "Name", "value": "John"},
+            {"field_id": "message", "label": "Message", "value": "just want to say hi"},
+            {"field_id": "name", "label": "Name", "value": "John"},
         ]
 
         response = self.submit_form(
@@ -820,7 +1097,7 @@ class TestMailSend(unittest.TestCase):
             },
         )
         transaction.commit()
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
         msg = self.mailhost.messages[0]
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
@@ -853,6 +1130,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "contact",
                         "field_type": "from",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -861,8 +1146,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "foo"},
                 ],
                 "block_id": "form-id",
@@ -876,8 +1165,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "foo@"},
                 ],
                 "block_id": "form-id",
@@ -891,8 +1184,12 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "foo@asd"},
                 ],
                 "block_id": "form-id",
@@ -903,7 +1200,7 @@ class TestMailSend(unittest.TestCase):
             response.json()["message"], 'Email not valid in "Email" field.'
         )
 
-    def test_submit_return_204_if_correct_email_in_email_field(
+    def test_submit_return_200_if_correct_email_in_email_field(
         self,
     ):
         """
@@ -921,6 +1218,14 @@ class TestMailSend(unittest.TestCase):
                         "field_id": "contact",
                         "field_type": "from",
                     },
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
                 ],
             },
         }
@@ -929,11 +1234,137 @@ class TestMailSend(unittest.TestCase):
         response = self.submit_form(
             data={
                 "data": [
-                    {"label": "Message", "value": "just want to say hi"},
-                    {"label": "Name", "value": "Smith"},
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "Smith"},
                     {"field_id": "contact", "label": "Email", "value": "foo@plone.org"},
                 ],
                 "block_id": "form-id",
             },
         )
-        self.assertEqual(response.status_code, 204)
+        self.assertEqual(response.status_code, 200)
+
+    def test_submit_return_200_with_submitted_data(self):
+        """
+        This is needed for confirm message
+        """
+        self.document.blocks = {
+            "form-id": {
+                "@type": "form",
+                "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
+            },
+        }
+        transaction.commit()
+
+        response = self.submit_form(
+            data={
+                "from": "john@doe.com",
+                "data": [
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "just want to say hi",
+                    },
+                    {"field_id": "name", "label": "Name", "value": "John"},
+                ],
+                "subject": "test subject",
+                "block_id": "form-id",
+            },
+        )
+        transaction.commit()
+        self.assertEqual(response.status_code, 200)
+        msg = self.mailhost.messages[0]
+        if isinstance(msg, bytes) and bytes is not str:
+            # Python 3 with Products.MailHost 4.10+
+            msg = msg.decode("utf-8")
+        self.assertIn("Subject: test subject", msg)
+        self.assertIn("From: john@doe.com", msg)
+        self.assertIn("To: site_addr@plone.com", msg)
+        self.assertIn("Reply-To: john@doe.com", msg)
+        self.assertIn("<strong>Message:</strong> just want to say hi", msg)
+        self.assertIn("<strong>Name:</strong> John", msg)
+
+    def test_cleanup_nasy_html_tags_with_portal_transforms(self):
+        """
+        This is needed for confirm message
+        """
+        self.document.blocks = {
+            "form-id": {
+                "@type": "form",
+                "send": ["recipient"],
+                "subblocks": [
+                    {
+                        "field_id": "message",
+                        "field_type": "text",
+                    },
+                    {
+                        "field_id": "name",
+                        "field_type": "text",
+                    },
+                ],
+            },
+        }
+        transaction.commit()
+
+        response = self.submit_form(
+            data={
+                "from": "john@doe.com",
+                "data": [
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "<b onmouseover=\"alert('XSS testing!')\">click here</b><p><i>keep tags</i></p>",
+                    },
+                    {
+                        "field_id": "name",
+                        "label": "Name",
+                        "value": "<script>alert(‘XSS’)</script> foo",
+                    },
+                ],
+                "subject": "test subject",
+                "block_id": "form-id",
+            },
+        )
+        transaction.commit()
+        self.assertEqual(response.status_code, 200)
+        res = response.json()
+
+        self.assertEqual(
+            res,
+            {
+                "data": [
+                    {
+                        "field_id": "message",
+                        "label": "Message",
+                        "value": "<b>click here</b><p><i>keep tags</i></p>",
+                    },
+                    {
+                        "field_id": "name",
+                        "label": "Name",
+                        "value": " foo",
+                    },
+                ]
+            },
+        )
+        msg = self.mailhost.messages[0]
+        if isinstance(msg, bytes) and bytes is not str:
+            # Python 3 with Products.MailHost 4.10+
+            msg = msg.decode("utf-8")
+        self.assertIn(
+            "<strong>Message:</strong> &lt;b&gt;click here&lt;/b&gt;&lt;p&gt;&lt=\r\n;i&gt;keep tags&lt;/i&gt;&lt;/p&gt;",
+            msg,
+        )
+        self.assertIn("<strong>Name:</strong>  foo", msg)